Don't assume that the git directory is .git
The git directory can be relocated by using an environment variable (GIT_DIR) or by making '.git' be a file with the path to the git directory. This set of changes leverages the 'git rev-parse' command to do the work of finding the git directory for ungit.

Fixes FredrikNoren/ungit/#1013.