Prevent re-render when text on a single line is selected
The change introduced [here](https://github.com/FredrikNoren/ungit/pull/1429#discussion_r495419717) (moving the click event to the row) requires a re-render when the row is clicked, and this causes selection issues when selecting text on a single line only.  This commit prevent the handler from running when patch mode is not enabled, and so makes selections from a single line possible at least in this case.