Re-render when patchLineList has been reset
When the edit state has been reset because all the patch line checkboxes have been cleared, `patchLineList` is reset to an empty list.  Comparing the before and after html strings is not adequate to detect this change, because `TextDiffViewModel.render` always renders with patchLineList initialized to all `true` values regardless.  Adding this check prevents the UI state from becoming inconsistent, while not rerendering on the `detectReActivity` event.